Soda 클래스 sugarContent Double로 수정

+++ VendingMachine/VendingMachine/Class/Soda/Coke.swift
@@ -11,7 +11,7 @@ import Foundation
 class Coke: Soda {
     private let caffeineContent: Bool
     
-    init(brand: String, capacity: Int, price: Int, name: String, date: Date, sugarContent: Int, caffeineContent: Bool) {
+    init(brand: String, capacity: Int, price: Int, name: String, date: Date, sugarContent: Double, caffeineContent: Bool) {
         self.caffeineContent = caffeineContent
         
         super.init(brand: brand, capacity: capacity, price: price, name: name, date: date, sugarContent: sugarContent)
